separated files widget and Window into separated files
